requirements.txt
fastapi==0.111.0             # Versão estável do FastAPI
uvicorn[standard]==0.30.1   # Versão estável do Uvicorn com extras para o servidor
sqlalchemy==2.0.30          # Versão estável e robusta do SQLAlchemy 2.0
pydantic==2.5.3             # Versão específica do Pydantic v2
pydantic-settings==2.0.3    # Versão específica de pydantic-settings compatível com Pydantic 2.5.3
PyMySQL==1.1.1              # Driver MySQL (PyMySQL, mais fácil de instalar)
python-jose[cryptography]==3.3.0 # Para JWT
passlib[bcrypt]==1.7.4      # Para hashing de senhas
openpyxl==3.1.2             # Para trabalhar com arquivos Excel (se ainda usar)


#python -m venv venv
#.\venv\Scripts\activate
#pip install -r requirements.txt
#uvicorn app.main:app --reload --port 8001 